refactor: move configurable to separate file
